+++ src/jace/__init__.py
@@ -9,12 +9,10 @@
 
 from __future__ import annotations
 
-import jax as _jax
-
 import jace.translator.primitive_translators as _  # noqa: F401  # Populate the internal registry.
 
 from .__about__ import __author__, __copyright__, __license__, __version__, __version_info__
-from .jax import grad, jacfwd, jacrev, jit
+from .api import grad, jacfwd, jacrev, jit
 
 
 __all__ = [
@@ -28,5 +26,3 @@ __all__ = [
     "__version__",
     "__version_info__",
 ]
-
-del _jax
